fix: add confirmation, add validate, and fix some ui on admin produk

+++ app/Http/Controllers/AdminProdukController.php
@@ -40,6 +40,20 @@ class AdminProdukController extends Controller
      */
     public function store(Request $request)
     {
+
+        $request->validate([
+            'nama' => 'required',
+            'kategoris' => 'required',
+            'gambar' => 'required',
+            'deskripsi' => 'required',
+            'harga' => 'required',
+            'komposisi' => 'required',
+            'karbo' => 'required',
+            'protein' => 'required',
+            'kalori' => 'required',
+            'serat' => 'required',
+        ]);
+
         if ($request->file('gambar')) {
             $gambar = $request->file('gambar')->store('images', 'public');
         }
@@ -62,7 +76,7 @@ class AdminProdukController extends Controller
             $produk->kategori()->attach($kats->id);
             $kats->save();
         }
-        return redirect()->route('product.index');
+        return redirect()->route('product.index')->with('success', 'Produk telah ditambahkan');;
     }
 
     /**
@@ -84,13 +98,31 @@ class AdminProdukController extends Controller
      */
     public function update(Request $request, $id)
     {
+
+        $request->validate([
+            'nama' => 'required',
+            'kategoris' => 'required',
+            'deskripsi' => 'required',
+            'harga' => 'required',
+            'komposisi' => 'required',
+            'karbo' => 'required',
+            'protein' => 'required',
+            'kalori' => 'required',
+            'serat' => 'required',
+        ]);
+
         $produk = Produk::findOrFail($id);
 
-        if ($produk->gambar && file_exists(storage_path('app/public/' . $produk->gambar))) {
-            Storage::delete('public/' . $produk->gambar);
-        }
-        $gambar = $request->file('gambar')->store('images', 'public');
-        $produk->gambar = $gambar;
+        if ($request->hasFile('gambar')) {
+            // Hapus gambar lama jika ada
+            if ($produk->gambar && file_exists(storage_path('app/public/' . $produk->gambar))) {
+                Storage::delete('public/' . $produk->gambar);
+            }
+        
+            // Simpan gambar baru
+            $gambar = $request->file('gambar')->store('images', 'public');
+            $produk->gambar = $gambar;
+        } 
 
         $produk->nama = $request->get('nama');
         $produk->deskripsi = $request->get('deskripsi');
@@ -105,7 +137,7 @@ class AdminProdukController extends Controller
         $kategoris = $request->get('kategoris');
         $produk->kategori()->sync($kategoris);
 
-        return redirect()->route('product.index');
+        return redirect()->route('product.index')->with('success', 'Produk telah diupdate');;
     }
 
     /**
@@ -121,6 +153,6 @@ class AdminProdukController extends Controller
         // Delete the produk
         $produk->delete();
 
-        return redirect()->route('product.index');
+        return redirect()->route('product.index')->with('success', 'Produk telah dihapus');;
     }
 }
